requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2024.6.1
ftfy==6.1.1
hf-xet==1.1.7
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.4
Jinja2==3.1.4
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.1
onnx==1.18.0
onnxruntime==1.16.3
opencv-python==4.8.1.78
packaging==25.0
pillow==11.0.0
protobuf==6.32.0
pycocotools==2.0.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2023.10.3
requests==2.28.1
safetensors==0.6.2
six==1.17.0
sympy==1.13.3
timm==0.9.12
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.12.2
urllib3==1.26.13
wcwidth==0.2.13
